Refactor(auth): extraer formularios y hooks de forgot/reset password

+++ app/Application/Usuario/Commands/Handlers/DestroyUsuarioHandler.php
@@ -3,7 +3,11 @@
 namespace App\Application\Usuario\Commands\Handlers;
 
 use App\Application\Usuario\Commands\DestroyUsuarioCommand;
+use App\Application\Usuario\Exceptions\CannotDeleteUsuarioException;
+use App\Application\Usuario\Exceptions\CannotFindUsuarioException;
+use App\Domains\Usuario\Aggregates\Usuario;
 use App\Domains\Usuario\Contracts\Repositories\UsuarioRepositoryContract;
+use App\Domains\Usuario\Enums\Roles;
 use App\Domains\Usuario\ValueObjects\UsuarioId;
 
 /**
@@ -23,6 +27,14 @@ final readonly class DestroyUsuarioHandler
 
     public function __invoke(DestroyUsuarioCommand $command): bool
     {
+        /** @var Usuario $usuario */
+        $usuario = $this->repository->findById(new UsuarioId($command->id));
+        if (!$usuario) {
+            throw new CannotFindUsuarioException();
+        }
+        if($usuario->isAdmin()){
+            throw new CannotDeleteUsuarioException('No se puede eliminar el usuario administrador');
+        }
         return $this->repository->destroy(new UsuarioId($command->id));
     }
 }
